+++ gax-httpjson/src/main/java/com/google/api/gax/httpjson/ManagedHttpJsonChannel.java
@@ -31,18 +31,11 @@ package com.google.api.gax.httpjson;
 
 import com.google.api.client.http.HttpTransport;
 import com.google.api.client.http.javanet.NetHttpTransport;
-import com.google.api.client.json.JsonFactory;
-import com.google.api.client.json.gson.GsonFactory;
-import com.google.api.core.ApiFuture;
 import com.google.api.core.BetaApi;
-import com.google.api.core.SettableApiFuture;
 import com.google.api.gax.core.BackgroundResource;
 import com.google.api.gax.core.InstantiatingExecutorProvider;
 import com.google.common.base.Preconditions;
-import com.google.common.collect.ImmutableList;
 import java.io.IOException;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
@@ -51,53 +44,33 @@ import javax.annotation.Nullable;
 /** Implementation of HttpJsonChannel which can issue http-json calls. */
 @BetaApi
 public class ManagedHttpJsonChannel implements HttpJsonChannel, BackgroundResource {
-  private static final JsonFactory JSON_FACTORY = GsonFactory.getDefaultInstance();
+
   private static final ExecutorService DEFAULT_EXECUTOR =
       InstantiatingExecutorProvider.newBuilder().build().getExecutor();
 
   private final Executor executor;
   private final String endpoint;
-  private final JsonFactory jsonFactory;
-  private final ImmutableList<HttpJsonHeaderEnhancer> headerEnhancers;
   private final HttpTransport httpTransport;
 
   private boolean isTransportShutdown;
 
+  protected ManagedHttpJsonChannel() {
+    this(null, null, null);
+  }
+
   private ManagedHttpJsonChannel(
-      Executor executor,
-      String endpoint,
-      JsonFactory jsonFactory,
-      List<HttpJsonHeaderEnhancer> headerEnhancers,
-      @Nullable HttpTransport httpTransport) {
+      Executor executor, String endpoint, @Nullable HttpTransport httpTransport) {
     this.executor = executor;
     this.endpoint = endpoint;
-    this.jsonFactory = jsonFactory;
-    this.headerEnhancers = ImmutableList.copyOf(headerEnhancers);
     this.httpTransport = httpTransport == null ? new NetHttpTransport() : httpTransport;
   }
 
   @Override
-  public <ResponseT, RequestT> ApiFuture<ResponseT> issueFutureUnaryCall(
-      HttpJsonCallOptions callOptions,
-      RequestT request,
-      ApiMethodDescriptor<RequestT, ResponseT> methodDescriptor) {
-    final SettableApiFuture<ResponseT> responseFuture = SettableApiFuture.create();
+  public <RequestT, ResponseT> HttpJsonClientCall<RequestT, ResponseT> newCall(
+      ApiMethodDescriptor<RequestT, ResponseT> methodDescriptor, HttpJsonCallOptions callOptions) {
 
-    HttpRequestRunnable<RequestT, ResponseT> runnable =
-        HttpRequestRunnable.<RequestT, ResponseT>newBuilder()
-            .setResponseFuture(responseFuture)
-            .setApiMethodDescriptor(methodDescriptor)
-            .setHeaderEnhancers(headerEnhancers)
-            .setHttpJsonCallOptions(callOptions)
-            .setHttpTransport(httpTransport)
-            .setJsonFactory(jsonFactory)
-            .setRequest(request)
-            .setEndpoint(endpoint)
-            .build();
-
-    executor.execute(runnable);
-
-    return responseFuture;
+    return new HttpJsonClientCallImpl<>(
+        methodDescriptor, endpoint, callOptions, httpTransport, executor);
   }
 
   @Override
@@ -138,16 +111,13 @@ public class ManagedHttpJsonChannel implements HttpJsonChannel, BackgroundResour
   public void close() {}
 
   public static Builder newBuilder() {
-    return new Builder()
-        .setHeaderEnhancers(new LinkedList<HttpJsonHeaderEnhancer>())
-        .setExecutor(DEFAULT_EXECUTOR);
+    return new Builder().setExecutor(DEFAULT_EXECUTOR);
   }
 
   public static class Builder {
+
     private Executor executor;
     private String endpoint;
-    private JsonFactory jsonFactory = JSON_FACTORY;
-    private List<HttpJsonHeaderEnhancer> headerEnhancers;
     private HttpTransport httpTransport;
 
     private Builder() {}
@@ -162,11 +132,6 @@ public class ManagedHttpJsonChannel implements HttpJsonChannel, BackgroundResour
       return this;
     }
 
-    public Builder setHeaderEnhancers(List<HttpJsonHeaderEnhancer> headerEnhancers) {
-      this.headerEnhancers = headerEnhancers;
-      return this;
-    }
-
     public Builder setHttpTransport(HttpTransport httpTransport) {
       this.httpTransport = httpTransport;
       return this;
@@ -174,8 +139,9 @@ public class ManagedHttpJsonChannel implements HttpJsonChannel, BackgroundResour
 
     public ManagedHttpJsonChannel build() {
       Preconditions.checkNotNull(endpoint);
+
       return new ManagedHttpJsonChannel(
-          executor, endpoint, jsonFactory, headerEnhancers, httpTransport);
+          executor, endpoint, httpTransport == null ? new NetHttpTransport() : httpTransport);
     }
   }
 }